requirements.txt
Flask==2.1.2
Flask-Cors==3.0.10
openai==0.27.0
python-dotenv==0.20.0
gunicorn==20.1.0
